retificação ok

+++ src/main/java/com/virtus/domain/dto/request/ProductElementRequestDTO.java
@@ -15,8 +15,8 @@ public class ProductElementRequestDTO extends BaseRequestDTO {
     private Integer componenteId;
     private Integer tipoNotaId;
     private Integer elementoId;
-    private Integer peso;
-    private Integer nota;
+    private Double peso;
+    private Double nota;
     private String motivacao;
 
 }
